Makefile: reintroduce format target, shuffle SRCS variables
To make the format target easier, we can hook into the $(SRCS) variable
before we add on the external source files, so we definitely format all
files that we would normally compile.

Headers are located via find -- a reasonable stopgap for now, but should
be improved in the future somehow.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>